+++ src/client/app/app.module.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
@@ -8,17 +8,20 @@
 
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { HttpModule }    from '@angular/http';
+import './rxjs-extensions';
 
-import { AppComponent }       from './app.component';
-import { DashboardComponent } from './dashboard/index';
-import { LoginComponent }     from './login/index';
-import { SettingsComponent }  from './settings/index';
-import { ToolbarComponent }   from './shared/index';
-import { routing }            from './app.routing';
+import { ApiIrcService, SidenavComponent, ToolbarComponent } from './shared/index';
+import { AppComponent }                                      from './app.component';
+import { DashboardComponent }                                from './dashboard/index';
+import { LoginComponent }                                    from './login/index';
+import { SettingsComponent }                                 from './settings/index';
+import { routing }                                           from './app.routing';
 
 @NgModule({
     imports: [
         BrowserModule,
+        HttpModule,
         routing,
     ],
     declarations: [
@@ -26,7 +29,11 @@ import { routing }            from './app.routing';
         DashboardComponent,
         LoginComponent,
         SettingsComponent,
+        SidenavComponent,
         ToolbarComponent,
+    ],
+    providers: [
+        ApiIrcService,
     ],
     bootstrap: [ AppComponent ],
 })