+++ src/client/app/app.routing.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
@@ -16,6 +16,11 @@ import { SettingsComponent }  from './settings/index';
 const appRoutes: Routes = [
     {
         path: '',
+        redirectTo: '/dashboard',
+        pathMatch: 'full',
+    },
+    {
+        path: 'dashboard',
         component: DashboardComponent,
     },
     {