+++ src/client/app/shared/api/index.ts
@@ -6,4 +6,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from './toolbar.component';
+export * from './api-irc.service';