+++ src/client/app/shared/index.ts
@@ -1,9 +1,11 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+export * from './api/index';
+export * from './sidenav/index';
 export * from './toolbar/index';