+++ src/client/app/shared/toolbar/toolbar.component.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
@@ -16,4 +16,8 @@ import { Component, Input } from '@angular/core';
 export class ToolbarComponent {
     @Input()
     isAuthenticated: boolean;
+
+    logout(): void {
+        window.location.href = '/logout';
+    }
 }