+++ src/server/plugin/hookCollection.ts
@@ -1,13 +1,13 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import * as api from 'twitchr-plugin-api';
-import {Plugin} from './pluginManager';
+import * as api   from 'twitchr-plugin-api';
+import { Plugin } from './pluginManager';
 
 export class HookCollection {
     private joinHooks: Array<api.PluginHook<api.IrcJoin>> = [];