+++ src/server/plugin/ircContext.ts
@@ -1,13 +1,13 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import * as api from 'twitchr-plugin-api';
-import {Client} from 'irc';
+import * as api   from 'twitchr-plugin-api';
+import { Client } from 'irc';
 
 export class IrcContext<T extends api.IrcEvent> implements api.IrcContext<T> {
     private channel: string;