+++ src/server/plugin/ircStore.ts
@@ -1,21 +1,21 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {IrcClient} from './ircClient';
+import { IrcClient } from './ircClient';
 
 interface IrcStore {
     [name: string]: IrcClient;
 }
 
-let store: IrcStore = {};
+const store: IrcStore = {};
 
 export function deleteClient(name: string): void {
-    let client: IrcClient = store[name];
+    const client: IrcClient = store[name];
     if (client) {
         client.stop(() => delete store[name]);
     }
@@ -25,6 +25,11 @@ export function getClient(name: string): IrcClient {
     return store[name];
 }
 
-export function putClient(name: string, client: IrcClient): void {
-    store[name] = client;
+export function putClient(name: string, token: string): IrcClient {
+    if (!store[name]) {
+        const client: IrcClient = new IrcClient(name, token);
+        store[name] = client.init();
+    }
+
+    return store[name];
 }