+++ src/server/routes/auth.ts
@@ -1,23 +1,23 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import * as express from 'express';
-import {authenticate} from 'passport';
-import {rejectAuthenticated} from '../util/auth';
+import * as express               from 'express';
+import { authenticate }           from 'passport';
+import { requireUnauthenticated } from '../util/auth';
 
 const router: express.Router = express.Router();
 
 router.get('/',
-    rejectAuthenticated,
+    requireUnauthenticated,
     authenticate('twitch'));
 
 router.get('/callback',
-    rejectAuthenticated,
+    requireUnauthenticated,
     authenticate('twitch', {
         failureRedirect: '/login',
         successRedirect: '/api/irc',