+++ src/server/routes/core.ts
@@ -1,19 +1,20 @@
 /**
  * @license
- * Copyright (C) 2016  Jonas Bürkel
+ * Copyright (C) 2017  Jonas Bürkel
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 import * as express from 'express';
-import {deleteClient} from '../plugin/ircStore';
-import {rejectAuthenticated, requireAuthenticated} from '../util/auth';
-import {resolve} from 'app-root-path';
+import { resolve }  from 'app-root-path';
+
+import { deleteClient }                                 from '../plugin/ircStore';
+import { requireAuthenticated, requireUnauthenticated } from '../util/auth';
 
 const router: express.Router = express.Router();
 
-router.get('/login', rejectAuthenticated, (req: express.Request, res: express.Response) => {
+router.get('/login', requireUnauthenticated, (req: express.Request, res: express.Response) => {
     res.sendFile(resolve('./dist/client/index.html'));
 });